Fix compiler warning due to sign comparison
When compiling with -Werror, this causes the build to fail.

Warning was:

| yamux_frame.cpp:103:28: error: comparison of integers of different signs: 'gsl::span::index_type' (aka 'int') and 'const uint32_t' (aka 'const unsigned int') [-Werror,-Wsign-compare]
|     if (frame_bytes.size() < YamuxFrame::kHeaderLength) {
|         ~~~~~~~~~~~~~~~~~~ ^ ~~~~~~~~~~~~~~~~~~~~~~~~~

+++ src/muxer/yamux/yamux_frame.cpp
@@ -100,7 +100,7 @@ namespace libp2p::connection {
   }
 
   boost::optional<YamuxFrame> parseFrame(gsl::span<const uint8_t> frame_bytes) {
-    if (frame_bytes.size() < YamuxFrame::kHeaderLength) {
+    if (frame_bytes.size() < static_cast<int>(YamuxFrame::kHeaderLength)) {
       return {};
     }
